+++ gym-system/src/main/java/com/gym/courseManagement/mapper/CourseReservationMapper.java
@@ -59,5 +59,5 @@ public interface CourseReservationMapper
      */
     public int deleteCourseReservationByReservationIds(Long[] reservationIds);
 
-    public int countReservationByTrainerId(Long trainerId);
+    public Long countReservationByTrainerId(Long trainerId);
 }
