+++ gym-system/src/main/java/com/gym/courseManagement/service/ICourseService.java
@@ -15,7 +15,7 @@ public interface ICourseService
     /**
      * trainerId计算课程数量
      */
-    int countCourseByTrainerId(Long trainerId);
+    Long countCourseByTrainerId(Long trainerId);
 
     /**
      * 通过courseId使得课程预约人数+1
